# Support non-uniform batch entries