--- a/Makefile
+++ b/Makefile
@@ -13,7 +13,6 @@
 #
 # Makefile for jdbi build targets
 #
-
 SHELL = /bin/sh
 .SUFFIXES:
 .PHONY: help clean install install-nodocker docs tests tests-nodocker publish-docs deploy release
@@ -56,10 +55,10 @@ release:
 help:
 	@echo " * clean            - clean local build tree"
 	@echo " * install          - builds and installs the current version in the local repository"
-	@echo " * install-nodocker - same as `install`, but skip all tests that require a local docker installation"
+	@echo " * install-nodocker - same as 'install', but skip all tests that require a local docker installation"
 	@echo " * docs             - build up-to-date documentation in docs/target/generated-docs/"
 	@echo " * tests            - run all unit and integration tests"
-	@echo " * tests-nodocker   - same as `tests`, but skip all tests that require a local docker installation"
+	@echo " * tests-nodocker   - same as 'tests', but skip all tests that require a local docker installation"
 	@echo " *"
 	@echo " ***********************************************************************"
 	@echo " *"