--- a/core/src/main/java/org/jdbi/v3/core/statement/PreparedBatch.java
+++ b/core/src/main/java/org/jdbi/v3/core/statement/PreparedBatch.java
@@ -66,7 +66,7 @@ public class PreparedBatch extends SqlStatement<PreparedBatch> implements Result
 
     public PreparedBatch(Handle handle, CharSequence sql) {
         super(handle, sql);
-        getContext().setBinding(new PreparedBinding(getContext()));
+        getContext().setBinding(new PreparedBinding(this, getContext()));
     }
 
     /**
@@ -246,7 +246,7 @@ public class PreparedBatch extends SqlStatement<PreparedBatch> implements Result
             beforeBinding();
 
             try {
-                ArgumentBinder binder = new ArgumentBinder.Prepared(this, parsedParameters, bindings.get(0));
+                ArgumentBinder binder = new ArgumentBinder(stmt, ctx, parsedParameters);
                 for (Binding binding : bindings) {
                     ctx.setBinding(binding);
                     binder.bind(binding);
@@ -263,7 +263,7 @@ public class PreparedBatch extends SqlStatement<PreparedBatch> implements Result
 
                 afterExecution();
 
-                ctx.setBinding(new PreparedBinding(ctx));
+                ctx.setBinding(new PreparedBinding(this, ctx));
 
                 return new ExecutedBatch(stmt, modifiedRows);
             } catch (SQLException e) {
@@ -285,7 +285,7 @@ public class PreparedBatch extends SqlStatement<PreparedBatch> implements Result
                     + "- call add() *after* setting batch parameters");
         }
         bindings.add(currentBinding);
-        getContext().setBinding(new PreparedBinding(getContext()));
+        getContext().setBinding(new PreparedBinding(this, getContext()));
         return this;
     }