… CONTRIBUTORS.md
… README.md
… RELEASE_NOTES.md
… bom/pom.xml
… core/src/test/java/org/jdbi/v3/core/ArgumentBinderPerf.java
… core/src/test/java/org/jdbi/v3/core/mapper/reflect/BeanMapperNestedTest.java
… core/src/test/java/org/jdbi/v3/core/mapper/reflect/BeanMapperPropagateNullableTest.java
… core/src/test/java/org/jdbi/v3/core/mapper/reflect/BeanMapperTest.java
… core/src/test/java/org/jdbi/v3/core/statement/TestArgumentBinder.java
… core/src/test/java/org/jdbi/v3/core/statement/TestScript.java
… docs/src/adoc/index.adoc
… internal/README.md
… pom.xml