Closes #2114 #2110 #2078 #1968